Display a 16x16 pixel grid as a class object
Define values for the width, the height, and the unit of mesurement
for each cell of the grid.

Use the 'class' library to represent the grid object as code, and
place all of its properties and functions in a separate file.

Draw the aforementioned grid to cover the entire screen.

Use for debugging.