Get user input to increase the chain size
Take the input from the user to set a growth flag to true. Set that
flag back to false right after a new square is added to the chain.

If the structure has a single square, and the growth flag is set to
true, the chain's size is increased, once, by 1 as well.

Change the head's color while the chain is growing, to display visual
feedback.

Update the chain's state every second.