Get user input to increase the chain velocity
Take the input from the user to set a momentum percentage to 25%. Use
that value to reduce the structure's refresh rate by that amount.

Ensure the refresh rate isn't so low that the game becomes unplayable.

Update the chain's state, depending on the amount of seconds specified
by the structure's refresh rate.